backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
httpx==0.25.2  # For async HTTP requests to other services
redis==5.0.1
psycopg2-binary==2.9.9  # For Postgres
sqlalchemy==2.0.23
alembic==1.13.1  # For database migrations
python-jose[cryptography]==3.3.0  # For JWT
passlib[bcrypt]==1.7.4  # For password hashing
python-multipart==0.0.6  # For form data